requirements.txt
pyrofork
tgcrypto
openai
python-dotenv
aiohttp[speedups]
beanie
apscheduler
tiktoken